Revert "Adjust type in arrays to save unused space"
This reverts commit 5f9e5fc9b8c7e1b8a094865ba13111a9a139446e.

 vitamin-caig commented on 5f9e5fc 2 hours ago
 Unaligned access is VERY slow for ARM/MIPS platforms (ie android), so I
 suggest you to keep best-match int type as it was earlier. I have improved
 YM2203 emulation performance up to 20% just by changing char/short 'economy'
 to int:)

The values are rarely used, but yes, aligned access is nicer.


Conflicts:
	src/lfo.c

+++ src/lfo.c
@@ -11,7 +11,7 @@
 
 #define WAVEFORM_SIZE 64
 
-static const short sine_wave[WAVEFORM_SIZE] = {
+static const int sine_wave[WAVEFORM_SIZE] = {
 	   0,  24,  49,  74,  97, 120, 141, 161, 180, 197, 212, 224,
 	 235, 244, 250, 253, 255, 253, 250, 244, 235, 224, 212, 197,
 	 180, 161, 141, 120,  97,  74,  49,  24,   0, -24, -49, -74,
